moved getmatch and it's unit test to public/js
